+++ noodles-bgzf/src/lib.rs
@@ -35,7 +35,7 @@
 //! ```
 
 #[cfg(feature = "async")]
-mod r#async;
+pub mod r#async;
 
 mod block;
 mod gz;